backend/requirements.txt
django>=5.1,<5.2
django-ninja>=1.3,<2.0
django-allauth>=65.0,<66.0
celery>=5.4,<5.5
django-celery-beat>=2.7,<2.8
redis>=5.0,<6.0
psycopg[binary]>=3.2,<4.0
dj-database-url>=2.2,<3.0
django-cors-headers>=4.4,<5.0
django-oauth-toolkit>=3.2,<4.0
python-decouple>=3.8,<4.0
gunicorn>=22.0,<23.0
whitenoise>=6.8,<7.0
requests>=2.31,<3.0
resend>=2.0,<3.0
cryptography>=46.0.6,<47.0
httpx>=0.27,<1.0
beautifulsoup4>=4.12,<5.0
lxml>=5.3,<6.0
PyJWT>=2.8,<3.0
PyYAML>=6.0,<7.0

# Cloud Tasks (async task dispatch in production)
google-cloud-tasks>=2.16,<3.0

# Capture AI extraction pipeline
google-auth>=2.36.0,<3.0.0
google-genai>=1.60.0,<2.0.0
anthropic[vertex]>=0.52.0,<1.0.0

# Dev / CI
ruff>=0.6,<1.0
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0
factory-boy>=3.3,<4.0
